+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Support for number and string fields in object-type arguments of `create()`.
+
 ## v1.2.1 - 2021-05-18
 
 ### Maintenance